% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_even_design.R
\name{even_even_design}
\alias{even_even_design}
\title{Constructs an orthogonal design for two types of even-level factors using
a Hadamard matrix. The first column of the Hadamard matrix is removed,
and scaled copies of the resulting matrix are combined to obtain the
required factor levels.}
\usage{
even_even_design(a, b, m)
}
\arguments{
\item{a}{Number of levels for the first factor. Must be even.}

\item{b}{Number of levels for the second factor. Must be even.}

\item{m}{Order of the Hadamard matrix.Must be a multiple of 4. The constructed design contains
\code{m - 1} factors of each of the two factor types.}
}
\value{
A matrix containing the constructed design. The first
\code{m - 1} columns correspond to the first type factor, and the remaining
\code{m - 1} columns correspond to the second type factor.
}
\description{
Constructs an orthogonal design for two types of even-level factors using
a Hadamard matrix. The first column of the Hadamard matrix is removed,
and scaled copies of the resulting matrix are combined to obtain the
required factor levels.
}
\keyword{internal}
