% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{generateGF}
\alias{generateGF}
\title{Generate Elements of a Finite Field
Generates the coefficient-vector representations of all
elements of GF(p^m).}
\usage{
generateGF(p, m)
}
\arguments{
\item{p}{A prime number.}

\item{m}{A positive integer specifying the field extension degree.}
}
\value{
A matrix containing all elements of GF(p^m).
}
\description{
Generate Elements of a Finite Field
Generates the coefficient-vector representations of all
elements of GF(p^m).
}
\keyword{internal}
