% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{gfAdd}
\alias{gfAdd}
\title{Finite-Field Addition
Adds two finite-field elements componentwise modulo p.}
\usage{
gfAdd(a, b, p)
}
\arguments{
\item{a}{A finite-field element.}

\item{b}{A finite-field element.}

\item{p}{A prime number.}
}
\value{
The sum of the two elements modulo p.
}
\description{
Finite-Field Addition
Adds two finite-field elements componentwise modulo p.
}
\keyword{internal}
