% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{gfMultiply}
\alias{gfMultiply}
\title{Finite-Field Multiplication
Multiplies two elements of GF(p^m) using polynomial
multiplication and reduction.}
\usage{
gfMultiply(a, b, p, modulus)
}
\arguments{
\item{a}{A finite-field element.}

\item{b}{A finite-field element.}

\item{p}{A prime number.}

\item{modulus}{The irreducible polynomial used to define the field.}
}
\value{
The product of the two finite-field elements.
}
\description{
Finite-Field Multiplication
Multiplies two elements of GF(p^m) using polynomial
multiplication and reduction.
}
\keyword{internal}
