% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{gfSubtract}
\alias{gfSubtract}
\title{Finite-Field Subtractio
Subtracts two finite-field elements componentwise modulo p.}
\usage{
gfSubtract(a, b, p)
}
\arguments{
\item{a}{A finite-field element.}

\item{b}{A finite-field element.}

\item{p}{A prime number.}
}
\value{
The difference of the two elements modulo p.
}
\description{
Finite-Field Subtractio
Subtracts two finite-field elements componentwise modulo p.
}
\keyword{internal}
