% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{irreduciblePolynomial}
\alias{irreduciblePolynomial}
\title{Irreducible Polynomial
Returns a predefined irreducible polynomial used for
constructing the finite field GF(p^m).}
\usage{
irreduciblePolynomial(p, m)
}
\arguments{
\item{p}{A prime number.}

\item{m}{A positive integer specifying the field extension degree.}
}
\value{
A numeric vector containing the polynomial coefficients.
}
\description{
Irreducible Polynomial
Returns a predefined irreducible polynomial used for
constructing the finite field GF(p^m).
}
\keyword{internal}
