% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeFunctions.R
\name{is_prime}
\alias{is_prime}
\title{Check for a Prime Number.
Checks whether a given number is prime.}
\usage{
is_prime(q)
}
\arguments{
\item{q}{A numeric value to be checked.}
}
\value{
TRUE if q is prime and FALSE otherwise.
}
\description{
Check for a Prime Number.
Checks whether a given number is prime.
}
\keyword{internal}
