% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeFunctions.R
\name{is_prime_power}
\alias{is_prime_power}
\title{Check for a Prime Power.
Checks whether q is a prime or a power of a prime number.}
\usage{
is_prime_power(q)
}
\arguments{
\item{q}{A positive integer to be checked.}
}
\value{
TRUE if q is a prime or prime power and FALSE otherwise.
}
\description{
Check for a Prime Power.
Checks whether q is a prime or a power of a prime number.
}
\keyword{internal}
