% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{jacobsthalPrime}
\alias{jacobsthalPrime}
\title{Jacobsthal Matrix for a Prime
Constructs a Jacobsthal matrix when q is prime using
the quadratic character modulo q.}
\usage{
jacobsthalPrime(q)
}
\arguments{
\item{q}{A prime number specifying the order of the
Jacobsthal matrix.}
}
\value{
A numeric Jacobsthal matrix of order q.
}
\description{
Jacobsthal Matrix for a Prime
Constructs a Jacobsthal matrix when q is prime using
the quadratic character modulo q.
}
\keyword{internal}
