% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{jacobsthalPrimePower}
\alias{jacobsthalPrimePower}
\title{Jacobsthal Matrix for a Prime Power
Constructs a Jacobsthal matrix over the finite field GF(p^m)
when q = p^m is an odd prime power.}
\usage{
jacobsthalPrimePower(q)
}
\arguments{
\item{q}{An odd prime power specifying the order of the
Jacobsthal matrix.}
}
\value{
A numeric Jacobsthal matrix of order q.
}
\description{
Jacobsthal Matrix for a Prime Power
Constructs a Jacobsthal matrix over the finite field GF(p^m)
when q = p^m is an odd prime power.
}
\keyword{internal}
