% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_odd_design.R
\name{odd_even_design}
\alias{odd_even_design}
\title{Constructs an orthogonal design for an odd-level factor and an
even-level factor using a conference matrix and a Hadamard matrix,
respectively.}
\usage{
odd_even_design(a, b, m)
}
\arguments{
\item{a}{Number of levels for the odd-level factor.}

\item{b}{Number of levels for the even-level factor.}

\item{m}{Order of the conference and Hadamard matrices. Must be a
multiple of 4.The design contains\code{m} factors of the odd-level type and \code{m - 1} factors of the
even-level type.}
}
\value{
A matrix containing the constructed design. The first \code{m}
columns correspond to the odd-level factors and are constructed
using a conference matrix, while the remaining \code{m - 1} columns
correspond to the even-level factors and are constructed using
a Hadamard matrix with its first column removed.
}
\description{
Constructs an orthogonal design for an odd-level factor and an
even-level factor using a conference matrix and a Hadamard matrix,
respectively.
}
\keyword{internal}
