% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_odd_design.R
\name{odd_odd_design}
\alias{odd_odd_design}
\title{Constructs an orthogonal design for two types of odd-level factors using
conference matrices.}
\usage{
odd_odd_design(a, b, m)
}
\arguments{
\item{a}{Number of levels for the first odd-level factor.}

\item{b}{Number of levels for the second odd-level factor.}

\item{m}{Order of the conference matrix. Must be a multiple of 4.The design contains
\code{m} factors of the first odd-level type and \code{m} factors of the second
odd-level type.}
}
\value{
A matrix containing the constructed design. The first \code{m}
columns correspond to the first odd-level factor and the remaining
\code{m} columns correspond to the second odd-level factor. Both sets of
columns are constructed using a conference matrix.
}
\description{
Constructs an orthogonal design for two types of odd-level factors using
conference matrices.
}
\keyword{internal}
