% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{paleyPrime}
\alias{paleyPrime}
\title{Paley Conference Matrix for a Prime
Constructs a Paley conference matrix of order n when
q = n - 1 is an odd prime.}
\usage{
paleyPrime(n)
}
\arguments{
\item{n}{A positive integer specifying the order of the
conference matrix.}
}
\value{
A numeric conference matrix of order n.
}
\description{
Paley Conference Matrix for a Prime
Constructs a Paley conference matrix of order n when
q = n - 1 is an odd prime.
}
\keyword{internal}
