% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{paleyPrimePower}
\alias{paleyPrimePower}
\title{Paley Conference Matrix for a Prime Power
Constructs a Paley conference matrix of order n when
q = n - 1 is an odd prime power.}
\usage{
paleyPrimePower(n)
}
\arguments{
\item{n}{A positive integer specifying the order of the
conference matrix.}
}
\value{
A numeric conference matrix of order n.
}
\description{
Paley Conference Matrix for a Prime Power
Constructs a Paley conference matrix of order n when
q = n - 1 is an odd prime power.
}
\keyword{internal}
