% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{polyMultiply}
\alias{polyMultiply}
\title{Polynomial Multiplication over GF(p
Multiplies two polynomials with coefficients in GF(p).}
\usage{
polyMultiply(a, b, p)
}
\arguments{
\item{a}{A vector containing the coefficients of the first polynomial.}

\item{b}{A vector containing the coefficients of the second polynomial.}

\item{p}{A prime number.}
}
\value{
A vector containing the coefficients of the product modulo p.
}
\description{
Polynomial Multiplication over GF(p
Multiplies two polynomials with coefficients in GF(p).
}
\keyword{internal}
