% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{polyReduce}
\alias{polyReduce}
\title{Polynomial Reduction over GF(p)
Reduces a polynomial modulo a specified irreducible polynomial
over GF(p).}
\usage{
polyReduce(poly, modulus, p)
}
\arguments{
\item{poly}{A vector containing polynomial coefficients.}

\item{modulus}{A vector containing the coefficients of the
irreducible polynomial.}

\item{p}{A prime number.}
}
\value{
The reduced polynomial as a coefficient vector.
}
\description{
Polynomial Reduction over GF(p)
Reduces a polynomial modulo a specified irreducible polynomial
over GF(p).
}
\keyword{internal}
