% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeFunctions.R
\name{primePowerInfo}
\alias{primePowerInfo}
\title{Identify a Prime Power
Determines whether q can be expressed as p^m, where p
is a prime number and m is a positive integer.}
\usage{
primePowerInfo(q)
}
\arguments{
\item{q}{A positive integer to be checked.}
}
\value{
A list containing p and m if q is a prime or
prime power; otherwise NULL.
}
\description{
Identify a Prime Power
Determines whether q can be expressed as p^m, where p
is a prime number and m is a positive integer.
}
\keyword{internal}
