% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoConference.R
\name{pseudoConference}
\alias{pseudoConference}
\title{Constructs a pseudo-conference matrix of order \code{n}
using a determinant-based search procedure.}
\usage{
pseudoConference(n, starts = 1000, max.iter = 100)
}
\arguments{
\item{n}{A positive integer specifying the order of the matrix.}

\item{starts}{Number of random starting matrices used in the
search. Default is 1000.}

\item{max.iter}{Maximum number of iterations for each starting
matrix. Default is 100.}
}
\value{
A list containing the pseudo-conference matrix and
the corresponding log-determinant value.
}
\description{
Constructs a pseudo-conference matrix of order \code{n}
using a determinant-based search procedure.
}
\details{
The matrix has zero diagonal elements and off-diagonal
elements equal to -1 or +1. The search selects a matrix
by maximizing the determinant criterion.
}
\examples{
result <- pseudoConference(6, starts = 10, max.iter = 10)
result$Matrix
result$LogDet
}
