% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{quadraticCharacterPrime}
\alias{quadraticCharacterPrime}
\title{Quadratic Character Modulo a Prime
Computes the quadratic character of an integer modulo a prime.
It returns 0 for zero, 1 for a quadratic residue, and -1 otherwise.}
\usage{
quadraticCharacterPrime(x, q, QR)
}
\arguments{
\item{x}{An integer.}

\item{q}{A prime number.}

\item{QR}{A vector containing the nonzero quadratic residues modulo q.}
}
\value{
The quadratic character: 0, 1, or -1.
}
\description{
Quadratic Character Modulo a Prime
Computes the quadratic character of an integer modulo a prime.
It returns 0 for zero, 1 for a quadratic residue, and -1 otherwise.
}
\keyword{internal}
