% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{quadraticResiduesGF}
\alias{quadraticResiduesGF}
\title{Quadratic Residues in a Finite Field
Finds the nonzero quadratic residues in GF(p^m).}
\usage{
quadraticResiduesGF(p, m)
}
\arguments{
\item{p}{A prime number.}

\item{m}{A positive integer specifying the field extension degree.}
}
\value{
A matrix containing the nonzero quadratic residues
of GF(p^m).
}
\description{
Quadratic Residues in a Finite Field
Finds the nonzero quadratic residues in GF(p^m).
}
\keyword{internal}
