% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleyConstruction.R
\name{quadraticResiduesPrime}
\alias{quadraticResiduesPrime}
\title{Quadratic Residues Modulo a Prime
Computes the nonzero quadratic residues modulo a prime number.}
\usage{
quadraticResiduesPrime(q)
}
\arguments{
\item{q}{A prime number.}
}
\value{
A vector containing the nonzero quadratic residues modulo q.
}
\description{
Quadratic Residues Modulo a Prime
Computes the nonzero quadratic residues modulo a prime number.
}
\keyword{internal}
