% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCD.R
\name{SCCD}
\alias{SCCD}
\title{Constructs an SCCD for a specified number of runs (n=2m).
The construction method is selected automatically according
to the number of runs.}
\usage{
SCCD(n)
}
\arguments{
\item{n}{A positive even integer specifying the number of runs.}
}
\value{
A numeric SCCD matrix containing m continuous and m-1
categorical factors.
}
\description{
Constructs an SCCD for a specified number of runs (n=2m).
The construction method is selected automatically according
to the number of runs.
}
\details{
Method I is used when n is divisible by 8.
Method II is used when n is divisible by 4 but not by 8.
Method III is used when n is even but not divisible by 4.
}
\examples{
SCCD(16)
}
