% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCD.R
\name{SCCD_method1}
\alias{SCCD_method1}
\title{Construct Saturated screening designs for continous and categorical factor (SCCD)
Using Method I
Constructs an SCCD when the number of runs (n=2m) is divisible by 8,
using a conference matrix and a Hadamard matrix.
The number of continuous factor is m and the number of categorical factors is m-1)}
\usage{
SCCD_method1(n)
}
\arguments{
\item{n}{A positive integer specifying the number of runs.}
}
\value{
A numeric SCCD matrix.
}
\description{
Construct Saturated screening designs for continous and categorical factor (SCCD)
Using Method I
Constructs an SCCD when the number of runs (n=2m) is divisible by 8,
using a conference matrix and a Hadamard matrix.
The number of continuous factor is m and the number of categorical factors is m-1)
}
\keyword{internal}
