% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCD.R
\name{SCCD_method2}
\alias{SCCD_method2}
\title{Construct SCCD Using Method II
Constructs an SCCD when the number of runs (n=2m) is divisible by 4
but not divisible by 8, using a conference matrix and a
D-optimal two-level design.
The number of continuous factor is m and number of categorical factors is m-1}
\usage{
SCCD_method2(n)
}
\arguments{
\item{n}{A positive integer specifying the number of runs.}
}
\value{
A numeric SCCD matrix.
}
\description{
Construct SCCD Using Method II
Constructs an SCCD when the number of runs (n=2m) is divisible by 4
but not divisible by 8, using a conference matrix and a
D-optimal two-level design.
The number of continuous factor is m and number of categorical factors is m-1
}
\keyword{internal}
