% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCD.R
\name{SCCD_method3}
\alias{SCCD_method3}
\title{Construct SCCD Using Method III
Constructs an SCCD when the number of runs (n=2m) is even but not
divisible by 4, using a pseudo-conference matrix and a
D-optimal two-level design.
The number of continuous factor is m and the number of categorical factors is m-1)}
\usage{
SCCD_method3(n, starts = 1000, max.iter = 100)
}
\arguments{
\item{n}{A positive integer specifying the number of runs.}

\item{starts}{Number of random starting matrices used in the
pseudo-conference matrix search. Default is 1000.}

\item{max.iter}{Maximum number of iterations for each starting
matrix. Default is 100.}
}
\value{
A numeric SCCD matrix.
}
\description{
Construct SCCD Using Method III
Constructs an SCCD when the number of runs (n=2m) is even but not
divisible by 4, using a pseudo-conference matrix and a
D-optimal two-level design.
The number of continuous factor is m and the number of categorical factors is m-1)
}
\keyword{internal}
