% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Performs a simulation study to evaluate the ability of a design
to identify active main effects using best-subsets regression
and the AICc criterion.}
\usage{
simulation(X, SNR, n.active, nrep = 1000)
}
\arguments{
\item{X}{A design matrix containing the mixed-level factors.}

\item{SNR}{A numeric value specifying the signal-to-noise ratio.}

\item{n.active}{An integer specifying the number of active factors.}

\item{nrep}{An integer specifying the number of simulation replications.
Default is 1000.}
}
\value{
A data frame containing the SNR, number of active factors,
power, and Type I error rate.
}
\description{
Performs a simulation study to evaluate the ability of a design
to identify active main effects using best-subsets regression
and the AICc criterion.
}
\examples{
\dontrun{
X <- matrix(rnorm(20 * 5), nrow = 20, ncol = 5)
simulation(X, SNR = 1, n.active = 2, nrep = 100)
}
}
