% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteField.R
\name{uniqueGFElements}
\alias{uniqueGFElements}
\title{Remove Duplicate Finite-Field Elements
Removes duplicate rows from a matrix of finite-field elements.}
\usage{
uniqueGFElements(M)
}
\arguments{
\item{M}{A matrix containing finite-field elements as rows.}
}
\value{
A matrix containing the unique finite-field elements.
}
\description{
Remove Duplicate Finite-Field Elements
Removes duplicate rows from a matrix of finite-field elements.
}
\keyword{internal}
